+++ build.gradle
@@ -1,5 +1,5 @@
 buildscript {
-	ext.kotlin_version = '1.3.61'
+	ext.kotlin_version = '1.3.72'
 	ext.junit_version = '5.5.2'
 	repositories {
 		mavenCentral()
@@ -32,26 +32,26 @@ compileTestKotlin {
 dependencies {
 	implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 
-	implementation 'io.javalin:javalin:3.5.0'
-	implementation 'org.slf4j:slf4j-simple:1.7.28'
-	implementation "com.fasterxml.jackson.core:jackson-databind:2.10.0.pr1"
-	implementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.10.0.pr1'
+	implementation 'io.javalin:javalin:3.+'
+	implementation 'org.slf4j:slf4j-simple:1.+'
+	implementation "com.fasterxml.jackson.core:jackson-databind:2.+"
+	implementation 'com.fasterxml.jackson.module:jackson-module-kotlin:2.+'
 
-	implementation 'org.hibernate.validator:hibernate-validator:6.1.0.Alpha6'
-	implementation 'org.glassfish:javax.el:3.0.1-b11'
+	implementation 'org.hibernate.validator:hibernate-validator:6.+'
+	implementation 'org.glassfish:javax.el:3.+'
 
 	implementation 'org.jetbrains.exposed:exposed:0.17.3'
-	implementation 'mysql:mysql-connector-java:8.0.17'
-    implementation 'org.litote.kmongo:kmongo:3.11.0'
+	implementation 'mysql:mysql-connector-java:8.+'
+    implementation 'org.litote.kmongo:kmongo:3.+'
 
 	testImplementation "org.junit.jupiter:junit-jupiter-engine:$junit_version"
-	testImplementation 'io.mockk:mockk:1.9.3'
-	testImplementation 'org.skyscreamer:jsonassert:1.5.0'
-	testImplementation 'com.tngtech.archunit:archunit-junit5-api:0.13.0'
-	testRuntime 'com.tngtech.archunit:archunit-junit5-engine:0.13.0'
+	testImplementation 'io.mockk:mockk:1.+'
+	testImplementation 'org.skyscreamer:jsonassert:1.+'
+	testImplementation 'com.tngtech.archunit:archunit-junit5-api:0.+'
+	testRuntime 'com.tngtech.archunit:archunit-junit5-engine:0.+'
 
-	testImplementation 'com.wix:wix-embedded-mysql:4.6.0'
-	testImplementation 'de.flapdoodle.embed:de.flapdoodle.embed.mongo:2.2.0'
+	testImplementation 'com.wix:wix-embedded-mysql:4.+'
+	testImplementation 'de.flapdoodle.embed:de.flapdoodle.embed.mongo:2.+'
 }
 
 test {
