+++ src/main/kotlin/users/web/CreateUser.kt
@@ -1,4 +1,4 @@
-package users.web.handlers
+package users.web
 
 import io.javalin.http.Context
 import io.javalin.http.Handler
