+++ src/test/kotlin/users/web/IntegrationTestWithMongoDB.kt
@@ -11,13 +11,12 @@ import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import users.WebAppConfig
 import users.domain.UserRepository
 import users.persistence.MongoDBUserRepository
 
 object IntegrationTestWithMongoDB {
 
-    private lateinit var webAppConfig: WebAppConfig
+    private lateinit var webApp: WebApp
     private lateinit var mongodExe: MongodExecutable
     private lateinit var mongod: MongodProcess
     private lateinit var userRepository: UserRepository
@@ -33,7 +32,7 @@ object IntegrationTestWithMongoDB {
         )
         mongod = mongodExe.start()
         userRepository = MongoDBUserRepository("localhost", 12345, "db123").apply { createSchema() }
-        webAppConfig = WebAppConfig(userRepository, 8081).apply { start() }
+        webApp = WebApp(userRepository, 8081).apply { start() }
     }
 
     @BeforeEach
@@ -54,7 +53,7 @@ object IntegrationTestWithMongoDB {
     @AfterAll
     @JvmStatic
     fun afterAll() {
-        webAppConfig.stop()
+        webApp.close()
         mongod.stop()
         mongodExe.stop()
     }
